python/requirements.txt
XTablesClient @ git+https://github.com/rokadias/XTABLES@d8f0f58757d7f7953e1342dd188f9880c2279d9d#subdirectory=src/main/python
